feat(api): :safety_vest: Add `StockHelpers` and `ValidationHelpers`
StockHelpers: Update the product quantity when the order is paid, and check for the quantity. Validation Helpers: Add validation for checking if entity is exisinting in database using the id. Add validation for checking if the user exist using id. Check validation if product exist and variant is exist within the product.